Extended bigint support to: * class Memory; some methods were still using int for sizes * class Vector, for size and indexing * all 1D "forall" macros and function templates   - HIP cannot launch kernels with >= 2^32 total threads   - CUDA seems to support kernel launches with >= 2^32 total threads * class DeviceTensor; individual dimensions still use int, however, the   total 1D index computation uses bigint * element and face geometric factors use bigint sizes for memory allocations
Left some FIXME comments to be addressed later.

+++ general/array.hpp
@@ -806,7 +806,7 @@ inline void Array<T>::Swap(Array &other)
 template <class T>
 inline void Array<T>::GrowSize(bigint minsize)
 {
-   const bigint nsize = std::max(bigint(minsize), 2 * data.Capacity());
+   const bigint nsize = std::max(minsize, 2 * data.Capacity());
    Memory<T> p(nsize, data.GetMemoryType());
    p.CopyFrom(data, size);
    p.UseDevice(data.UseDevice());
